% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_exact.R
\name{match_exact}
\alias{match_exact}
\alias{match_exact.character}
\title{Exact matcher}
\usage{
match_exact(x, ref = NULL, against = NULL)

\method{match_exact}{character}(x, ref = NULL, against = NULL)
}
\arguments{
\item{x}{Input species list, a character vector}

\item{ref}{(character) Reference taxon data.frame, or file path}

\item{against}{(character) What column to match against in data.frame.
Ignored if a vector given}
}
\description{
Exact matcher
}
\examples{
x <- c("Salmo eperlanus Linnaeus, 1758", 'Oncorhynchus clarkii', 'Salmo',
'Oncorhynchus clarkii', 'Salvelinus fontinalis', 'Salvelinus confluentus')
y <- system.file("examples", "worms_salmo.csv", package = "splister")
dat <- unique(read.csv(y, stringsAsFactors = FALSE))

# get exact matches
res <- match_exact(x, ref = dat, against = "scientificName")

# then move on to do fuzzy matching
match_fuzzy(x, dat, against = "scientificName")
}

