% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_fuzzy.R
\name{match_fuzzy}
\alias{match_fuzzy}
\alias{match_fuzzy.splist}
\title{Fuzzy matcher}
\usage{
match_fuzzy(x, ref = NULL, against = NULL, ...)

\method{match_fuzzy}{splist}(x, ref = NULL, against = NULL, ...)
}
\arguments{
\item{x}{Input species list, a character vector}

\item{ref}{Reference taxon data.frame, or file path}
}
\description{
Fuzzy matcher

match_fuzzy.character <- function(x, ref = NULL, against = NULL) {
  if (!is(ref, "data.frame")) ref <- readr::read_csv(ref)
  structure(unname(lapply(x, function(z) {
    ex <- ref[ref[[against]] %in% z, against]
    if (length(ex) == 0) {
      structure(list(z, NA), match = "no_exact_match")
    } else {
      structure(list(z, ex), match = "exact_match")
    }
  })), class = "splist", ref = what_ref(ref), against = against)
}
}
\examples{
x <- c("Salmo eperlanus Linnaeus, 1758", 'Oncorhynchus clarkii', 'Salmo',
'Oncorhynchus clarkii', 'Salvelinus fontinalis', 'Salvelinus confluentus')
y <- system.file("examples", "worms_salmo.csv", package = "splister")
dat <- unique(read.csv(y, stringsAsFactors = FALSE))

# get exact matches
res <- match_exact(x, ref = dat, against = "scientificName")

# then move on to do fuzzy matching
match_fuzzy(res)

# pipe, if you like
match_exact(x, ref = dat, against = "scientificName") \%>\% match_fuzzy

# read from file
x <- system.file("examples", "iucn_dat.csv", package = "splister")
x <- read.csv(x, stringsAsFactors = FALSE)[,-1]
y <- system.file("examples", "worms_sample.csv", package = "splister")
spp <- x$sciname[1:5000L]
(res <- match_exact(spp, ref = y, against = "scientificName"))
res2 <- match_fuzzy(res)
}

