% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_names.R
\name{pull_names}
\alias{pull_names}
\alias{pull_names.splist}
\title{Pull out names after matching is done}
\usage{
pull_names(x)

\method{pull_names}{splist}(x)
}
\arguments{
\item{x}{Input species list, a character vector}
}
\value{
Characater vector of names, original names for those not matched,
and replacment names for those matched
}
\description{
Pull out names after matching is done
}
\examples{
\dontrun{
x <- system.file("examples", "iucn_dat.csv", package = "splister")
x <- read.csv(x, stringsAsFactors = FALSE)[,-1]
y <- system.file("examples", "worms_sample.csv", package = "splister")
spp <- x$sciname[1:1000L]
res <- match_exact(spp, ref = y, against = "scientificName")
res2 <- match_fuzzy(res)
pull_names(res2)
}
}

