% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splister-package.R
\docType{package}
\name{splister-package}
\alias{splister}
\alias{splister-package}
\title{Match Species Lists Against Reference List}
\description{
Match Species Lists Against Reference List
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

