% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_names.R
\name{case_names}
\alias{case_names}
\title{Change case}
\usage{
case_names(x, what = "onlyfirst")
}
\arguments{
\item{x}{Input species list, a character vector}
}
\value{
Same as put in, either list, character vector or data.frame
}
\description{
Change case
}
\examples{
x <- c("Salmo eperlanus Linnaeus, 1758", 'Oncorhynchus clarkii', 'Salmo',
'Oncorhynchus clarkii', 'Salvelinus fontinalis', 'Salvelinus confluentus')

# capitalize first letter only
case_names(x)

# all lower case
case_names(x, 'lower')

# all upper case
case_names(x, 'upper')

# trim white space
x <- c("Salmo eperlanus Linnaeus, 1758  ", 'Oncorhynchus clarkii', 'Salmo ',
'Oncorhynchus clarkii ', '   Salvelinus fontinalis ', 'Salvelinus confluentus  ')
case_names(x, 'trim')

}

