% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_fuzzy.R
\name{match_fuzzy}
\alias{match_fuzzy}
\alias{match_fuzzy.character}
\alias{match_fuzzy.splist}
\title{Fuzzy matcher}
\usage{
match_fuzzy(x, ref = NULL, against = NULL, ...)

\method{match_fuzzy}{character}(x, ref = NULL, against = NULL, ...)

\method{match_fuzzy}{splist}(x, ref = NULL, against = NULL, ...)
}
\arguments{
\item{x}{Input species list, a character vector}

\item{ref}{Reference taxon data.frame, or file path}

\item{against}{(character) What column to match against in data.frame.
Ignored if a vector given}

\item{...}{Further args passed on to \code{\link{grep}}}
}
\description{
Fuzzy matcher
}
\examples{
\dontrun{
x <- c("Salmo eperlanus Linnaeus, 1758", 'Oncorhynchus clarkii', 'Salmo',
'Oncorhynchus clarkii', 'Salvelinus fontinalis', 'Salvelinus confluentus')
y <- system.file("examples", "worms_salmo.csv", package = "splister")
dat <- unique(read.csv(y, stringsAsFactors = FALSE))

# get exact matches
res <- match_exact(x, ref = dat, against = "scientificName")

# then move on to do fuzzy matching
match_fuzzy(res)

# pipe, if you like
match_exact(x, ref = dat, against = "scientificName") \%>\% match_fuzzy

# read from file
x <- system.file("examples", "iucn_dat.csv", package = "splister")
x <- read.csv(x, stringsAsFactors = FALSE)[,-1]
y <- system.file("examples", "worms_sample.csv", package = "splister")
spp <- x$sciname[1:5000L]
(res <- match_exact(spp, ref = y, against = "scientificName"))
res2 <- match_fuzzy(res)
}
}

