% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{matcher}
\alias{matcher}
\alias{matcher.character}
\title{Match spp list to reference list}
\usage{
matcher(spplist, ref, against = NULL)

\method{matcher}{character}(spplist, ref, against = NULL)
}
\arguments{
\item{spplist}{Input species list, a character vector}

\item{ref}{Reference taxon data.frame, or file path}

\item{against}{(character) What column to match against in data.frame.
Ignored if a vector given}
}
\description{
Match spp list to reference list
}
\examples{
x <- c('Salmo trutta', 'Oncorhynchus clarkii', 'Salmo Linnaeus',
'Oncorhynchus clarkii', 'Salvelinus fontinalis', 'Salvelinus confluentus')
y <- system.file("examples", "worms_salmo.csv", package = "splister")
dat <- unique(read.csv(y, stringsAsFactors = FALSE))
matcher(spplist = x, ref = dat, against = "scientificName")

# read from file
x <- system.file("examples", "iucn_dat.csv", package = "splister")
x <- read.csv(x, stringsAsFactors = FALSE)[,-1]
y <- system.file("examples", "worms_sample.csv", package = "splister")
spp <- x$sciname[1:200L]
res <- matcher(spplist = spp, ref = y, against = "scientificName")
unique(sapply(res, attr, "match"))
df <- data.frame(original = spp, taxon = unlist(res),
   match = sapply(res, attr, "match"))
df[df$match == "replace", ]
}

