% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{matcher}
\alias{matcher}
\alias{matcher.data.frame}
\title{Match spp list to reference list}
\usage{
matcher(spplist, ref)

\method{matcher}{data.frame}(spplist, ref)
}
\arguments{
\item{spplist}{Input species list, a vector}

\item{ref}{Reference taxon data.frame}
}
\description{
Match spp list to reference list
}
\examples{
x <- ""
y <- ""
matcher(x, y)
}

